+++ clients/admin-ui/src/features/dataset/DatasetCollectionView.tsx
@@ -47,9 +47,7 @@ const DatasetCollectionView = ({ fidesKey }: Props) => {
   const dispatch = useDispatch();
   const { dataset, isLoading } = useDataset(fidesKey);
   const activeCollectionIndex = useSelector(selectActiveCollectionIndex);
-  const [columns, setColumns] = useState<ColumnMetadata[]>(
-    ALL_COLUMNS.filter((c) => c.attribute !== "data_categories")
-  );
+  const [columns, setColumns] = useState<ColumnMetadata[]>(ALL_COLUMNS);
   const [isModifyingCollection, setIsModifyingCollection] = useState(false);
   const [isModifyingDataset, setIsModifyingDataset] = useState(false);